refactor docusaurus-utils hash fns
